+++ logs/orderbook_log_(06-04-2025_20:38:14).txt
@@ -1,0 +1,60 @@
+20:38:14 [INFO] Bid Book is empty.
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Ask Book is empty.
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Trade List are empty!
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=1, price: 15, qty=100, side=buy, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=2, price: 23.5, qty=100, side=buy, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=3, price: 25, qty=100, side=buy, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=4, price: 250, qty=100, side=buy, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=5, price: 23.5, qty=100, side=sell, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Trade recorded: <Trade (buy_order_id=4, sell_order_id=5, price=23.5, qty=100, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=6, price: 25, qty=100, side=sell, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Trade recorded: <Trade (buy_order_id=3, sell_order_id=6, price=25, qty=100, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=7, price: 50, qty=100, side=sell, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order limit: <LimitOrder id=8, price: 250, qty=100, side=sell, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Bid: 23.5
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Ask: 50
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Trade List:
+20:38:14 [INFO] index: 0, trade: <Trade (buy_order_id=4, sell_order_id=5, price=23.5, qty=100, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] index: 1, trade: <Trade (buy_order_id=3, sell_order_id=6, price=25, qty=100, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Ask Orders:
+20:38:14 [INFO] Order ID:8, Price: 250, Qty: 100
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Order ID:7, Price: 50, Qty: 100
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Bid Orders:
+20:38:14 [INFO] Order ID: 2, Price: 23.5, Qty: 100
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Order ID: 1, Price: 15, Qty: 100
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Placing order market: <MarketOrder id=9, qty=20, side=buy, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Trade recorded: <Trade (buy_order_id=9, sell_order_id=7, price=50, qty=20, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Average price for Market Order is: 50.0
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Matching Market Order: <MarketOrder id=9, qty=0, side=buy, asset=BTC-USD>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] Trade List:
+20:38:14 [INFO] index: 0, trade: <Trade (buy_order_id=4, sell_order_id=5, price=23.5, qty=100, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] index: 1, trade: <Trade (buy_order_id=3, sell_order_id=6, price=25, qty=100, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
+20:38:14 [INFO] index: 2, trade: <Trade (buy_order_id=9, sell_order_id=7, price=50, qty=20, asset=BTC-USD)>
+20:38:14 [INFO] ________________________________________________________________________________
