+++ utils/helpers.py
@@ -1,0 +1,4 @@
+from utils.logger import logger
+
+def print_line(length=80):
+    logger.info("_" * length)
